updates from eks testing

+++ 6-forkApplicationRepositories.sh
@@ -40,7 +40,9 @@ echo ""
 echo "*** NOTE: This will first delete the forked repos "
 echo "          in the the target github organization ***"
 echo "===================================================="
-read -rsp $'Press ctrl-c to abort. Press any key to continue...\n' -n1 key
+if [ "$1" == "skip" ]; then
+  read -rsp $'Press ctrl-c to abort. Press any key to continue...\n' -n1 key
+fi
 echo ""
 
 rm -rf repositories/
